osd/ReplicatedPG: explicitly remove objects when deleting them
In the EC case, we would rely on log processing of delete log entries
to add a remove operation to the local transaction. Make this explicit
in the transaction to begin with instead, so it works when log
processing does not handle the delete.
Note: this is not a cherry-pick because in master ec-overwrites rewrote
this code so it was not an issue.

Signed-off-by: Josh Durgin <jdurgin@redhat.com>

Resolves: rhbz#1452780