qa/standalone/ceph-helpers: disable core check
This fails for me because many unit tests generate cores naturally
(specifically, the unit tests have EXPECT_DEATH and those generate
a core each time).  This makes my local runs fail the standalone
task nondeterministically, depending on whether they run after
any of the death tests (the usually do in my case).

I'm not sure how to distinguish between the two... which makes me
inclined to just disable this for now so that the tests work.

Signed-off-by: Sage Weil <sage@redhat.com>