lockdep: prevent reuse of lockdep ids across lockdep_unregister_ceph_context
It's possible for the teardown of g_lockdep_ceph_ctx to occur, followed
by a new context being registered as the lockdep context. When that
occurs, we can end up reusing lock id's that were previously handed out
to consumers.

Make both the free_ids table, and the lock_refs map persistent across
lockdep_unregister_ceph_context and lockdep_register_ceph_context.

When lockdep_unregister is called, do the refcounting like we normally
would, but only clear out the lock_ids, lock_names, and follows_bt
hashes if the lockid is registered in the lock_names hash. In either
case, free up the id.

Finally, we do still need to handle the case where g_lockdep has gone
false even when there are outstanding references after the decrement.
Only log the message to the lock if that's not the case.

Tracker: http://tracker.ceph.com/issues/21512
Signed-off-by: Jeff Layton <jlayton@redhat.com>