mds/SnapServer: use recent_removed_snaps
We assume we will process OSDMaps frequently enough to catch a removed
snap via recent_removed_snaps.  We ideally would use the new_removed_snaps
for this purpose, but the MDS does not necessarily see every single OSDMap
(although it usually does).

This should be fine in practice because only a live MDS will be removing
snaps.  If the MDS is offline for a period of time (such that it does not
see OSDMaps) there would not be any new removed snaps for that period.

Signed-off-by: Sage Weil <sage@redhat.com>