osd/PG: _update_calc_stats: backfill targets do not always affect degraded
Our progress on a backfill target should only count toward degraded if
the target is needed in order for us to reach the target pool size.  If
we have remapped other complete copies, we can backfill and not be degraded
at all.  And we may be some combination of the two if there are multiple
backfill targets; in that case, we arbitrarily count the first target(s)
as the one within the target size set that counts against degraded and the
later ones do not affect degraded.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit 1cfffc2945a2eefc7f02fc04f8f4f67c0afd2ef3)