osd/PrimaryLogPG: clear DEGRADED at recovery completion even if more backfill
We may have log recovery *and* backfill to do, but cease to be degraded
as soon as the log recovery portion is done.  If that's the case, clear
the DEGRADED bit so that the PG state is not misleading.

Signed-off-by: Sage Weil <sage@redhat.com>
(cherry picked from commit f475ebbda8bd1e4ea9ef07920fcc1067e1891833)