mon/OSDMonitor: add option to fix up ruleset-* to crush-* for ec profiles
The jewel->luminous upgrade test will fail if we finish the upgrade while
a workload setting old-style ec profiles is running.  Add option to
automatically fix them up.  Warn to the cluster log when this happens.

For now, enable this option to ease upgrades and whitelist the warning.

Only include this option in luminous so that we implicitly sunset this
compatibility kludge immediately.

Fixes: http://tracker.ceph.com/issues/22128
Signed-off-by: Sage Weil <sage@redhat.com>