mon/Monitor: fix statfs handling before luminous switchover happens
After the mons are luminous but before we switch over to using the
MgrStatMonitor's new info, the version on mgrstat will generally be <<
than that of pgmon, and the client will send that version with the
request.  This means that the statfs message will perpetually appear to be
in the future and fail the is_readable() check.

Fix this with any ugly hack that resets the version to 1 if we haven't
completed the luminous upgrade yet.

Signed-off-by: Sage Weil <sage@redhat.com>