Store swift acls with keystone elements.
In swift, acls can contain keystone elements.  Keystone elements always
have project : user where project & user can be names (deprecated), ids,
or '*'.  In the master branch of ceph, these are stored as acl elements
with an empty name field and an id string that contains the element
string.  This is done with a bit of code that says "skipping silently".
(In jewel they really were being ignored silently.)

This code adds logic to emulate the same behavior in jewel.  This is
just the store logic for that, not the actual permissions logic.

The master branch also has logic to parse and store "referral" grants.
This commit does not include equivalent logic to do so, because the
supporting logic framework is not present.

This is part 1 of 4 fixes for tracker issue #22898.

Fixes: http://tracker.ceph.com/issues/22898
Resolves: rhbz#1516099

Signed-off-by: Marcus Watts <mwatts@redhat.com>
Conflicts:
 The master branch has entirely different code for this logic
 which already provides similar functionality.
(cherry picked from commit 71707dac181196116665e3190d1a75f2f8bdac0e)