client: allow getattr and ll_getattr to use implemented caps
caps_issued_mask is a function that we call to see if we have all of
the caps provided in a mask. Usually this ignores caps that have been
revoked by the MDS but that have not yet been released by the client.
These caps are still valid though -- nothing else can use them until
they've been released.

The getattr codepaths call caps_issued_mask only in order to determine
whether the client holds enough caps to fill out all of the requested
attributes. They don't take new references to the caps, so it should
be safe to use the "implemented" capset there.

This helps break a deadlock that can occur when the client holds a
delegation that has been recalled, but needs to satisfy a getattr
request before releasing it.

We may need to extend this to other codepaths as well, but it's
probably best to do that in other patches.

Tracker: http://tracker.ceph.com/issues/23028
Signed-off-by: Jeff Layton <jlayton@redhat.com>