PG: do not choose osds that are not up as async_recovery_targets
If we accept osds that are down as async_recovery_targets, we need to
ensure that they get a chance to become part of the acting set after
recovery is over.

However, when choose_acting() is called in the Recovered state, we set
restrict_to_up_acting=true, which does not allow the asychronously
recovered down osd to get back to the acting set.

Therefore, similar to backfill, do not allow osds that are not up to
become an async_recovery_target.

Signed-off-by: Neha Ojha <nojha@redhat.com>