mgr: add pg_state_ready for pgs show in unknown state
Pgs show the unknown state, when the mgr has just started, use the command ceph pg dump pgs_brief\pgs\...,
because the pgs state has not been reported from the Osd to the Mgr service,
we need to add this state for pgs show. in this way,our management system can judge whether to filter
the unknown pg state based on this flag.

Fixes: http://tracker.ceph.com/issues/25103
Signed-off-by: huanwen ren <ren.huanwen@zte.com.cn>