mgr/dashboard: Verify the RGW mail reset behavior
The current behavior when submitting an empty mail field in the RGW user form
that was previously filled was to replace it with "false".

This was a regression introduced through the use of "CdFormGroup" which had
a "_filterValue" method which converted empty strings into false, but it was
removed now.

Fixes: http://tracker.ceph.com/issues/26861
Signed-off-by: Stephan Müller <smueller@suse.com>
(cherry picked from commit a159f5a1be12403373e25d9bdbf7a405562da1d0)
Signed-off-by: Ernesto Puerta <epuertat@redhat.com>