mgr/ssh: rename ssh_config file system path option
this option specifies a path on the local file system where an ssh
config file is located. this is being renamed to avoid conflicts with
the ssh_config option in a subsequent patch that accepts the file
contents of an ssh_config to be specified without requiring it to be
saved to disk by the user.

Signed-off-by: Noah Watkins <noahwatkins@gmail.com>