mgr/dashboard/qa: Improve tasks.mgr.test_dashboard.TestDashboard.test_standby
The changes of this PR were done while trying to fix the failing test. The problem has been solved by another PR, but the changes are worth to be integrated because they help debugging and an additional test has been added (check if previously active manager is listed as standby).

Signed-off-by: Volker Theile <vtheile@suse.com>