rgw: auto-clean reshard queue entries for non-existent buckets
It is possible for a bucket to be added to the reshard queue and then
to be removed before its entry in the reshard queue is
processed. Prior to this pr, the entry would remain on the reshard
queue. With this pr, the entry will be automatically removed with
appropriate log entries.

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>