qa: Run flake8 on python2 and python3
To be able to catch problems with python2 *and* python3, run flake8
with both versions. From the flake8 homepage:

It is very important to install Flake8 on the correct version of
Python for your needs. If you want Flake8 to properly parse new
language features in Python 3.5 (for example), you need it to be
installed on 3.5 for Flake8 to understand those features. In many
ways, Flake8 is tied to the version of Python on which it runs.

Also fix the problems with python3 on the way.
Note: This requires now the six module for teuthology. But this is
already an install_require in teuthology itself.

Signed-off-by: Thomas Bechtold <tbechtold@suse.com>
(cherry picked from commit bdcc94a1d1f659b1524fdc6bb8bd1da6d38a30d2)

Conflicts:
	qa/CMakeLists.txt
	qa/tasks/ceph_manager.py
	qa/tasks/cephfs/xfstests_dev.py
	qa/tasks/cram.py
	qa/tasks/manypools.py
	qa/tasks/ragweed.py
	qa/tasks/s3tests.py
	qa/tasks/vstart_runner.py
	qa/tasks/workunit.py
	qa/tox.ini: trivial resolutions